services/message-service/requirements.txt
fastapi
uvicorn
azure-functions
